changed structure of files
