+++ controllers/data.js
@@ -1,17 +1,3 @@
-// require('./srvc.js');
-var app = angular.module('myApp', ['ui.router']);
-
-app.controller('mainCtrl', function($scope, $state) {
-    $scope.tabs = ['intro', 'first', 'second', 'third', 'fourth', 'fith', 'summary'];
-    $scope.currentIndex = 0;
-    $scope.nextq = function() {
-        $scope.currentIndex = $scope.currentIndex + 1;
-        var name = $scope.tabs[$scope.currentIndex];
-        $state.go(name);
-    }
-});
-
-
 angular.module("myApp").controller("newctrl", newController);
 
 newController.$inject = [
@@ -24,6 +10,9 @@ function newController($scope, CalculatorService) {
      $scope.answer = CalculatorService.square($scope.number);
    }
 }
+
+
+
 // app.controller("add", ['sharedList', function(sharedList) {
 //   var vm = this;
 //   vm.text = "";
@@ -63,10 +52,3 @@ function newController($scope, CalculatorService) {
 //      console.log(vm.answers);
 //    }
 // });
-
-// TODO:
-// - Unhide questions one by one (and hide previous one)
-// - Validations (if "next" clicked without choosing option)
-// - One dynamic question (question 4)
-// - Storing the chosen options
-// - Showing everything in summary
