+++ index.html
@@ -7,7 +7,8 @@
     <link type="text/css" rel="stylesheet" href="stylesheet.css" />
     <script src="https://ajax.googleapis.com/ajax/libs/angularjs/1.6.4/angular.min.js"></script>
     <script src="//unpkg.com/angular-ui-router/release/angular-ui-router.min.js"></script>
-    <script type="text/javascript" src="form.js"></script>
+    <script type="text/javascript" src="Controllers/form.js"></script>
+        <script type="text/javascript" src="data/form.js"></script>
     <script type="text/javascript" src="routing.js"></script>
     <script type="text/javascript" src='srvc.js'></script>
     <title>Webpage</title>
